Split testing file by api
